requirements.txt
goose3==3.1.16
loguru==0.7.0
nltk==3.8.1
trafilatura==1.6.1
newspaper3k==0.2.8
